backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI Services
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
ollama==0.1.6

# Voice Services
speechrecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
pyaudio==0.2.14

# Task & Calendar
python-dateutil==2.8.2
icalendar==5.0.11
croniter==2.0.1

# Web Search
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.3

# File Analysis
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
pandas==2.1.3

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
